:bug: Fixed permissions issue with acme.json file
The Traefik service was failing to start due to incorrect permissions on the acme.json file. The file permissions were too open, which caused Traefik to skip the ACME resolver.

The solution was to change the permissions of the acme.json file to 600, which is the level of security that Traefik requires. This was done by running a temporary Docker container with the volume mounted, and changing the permissions of the file from within the container.