:bug: Fix Bad Gateway error for localhost access
This commit fixes a 502 Bad Gateway error when trying to access the game service at https://localhost:9443/game-logic. The error was due to a misconfiguration in the Traefik labels for the game service.

The service could be accessed via the local IP (https://10.24.107.1:9443/game-logic), but not localhost. The changes in this commit ensure that the service can be accessed via both localhost and the local IP.

Additionally, this commit adds comments to the Traefik labels for the react-dev service, explaining their purpose.