:sparkles: Enable WebSocket connections from any host
This commit enables the front-end to establish WebSocket connections from any host, whether it's localhost or another computer on the same network. This is achieved by dynamically setting the WebSocket URL based on the hostname of the current page.

Previously, the WebSocket URL was hardcoded to 'localhost', which prevented connections from other computers. Now, the WebSocket URL is set to the hostname of the current page, allowing connections from any host.

This change improves the flexibility and usability of the application, as it can now be used on any computer in the same network.